pyproject.toml
[project]
name = "artrefsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bidict>=0.23.1",
    "certifi>=2026.1.4",
    "dacite>=1.9.2",
    "darkdetect>=0.8.0",
    "dataclasses-json>=0.6.7",
    "diskcache>=5.6.3",
    "dnspython[doh]>=2.6.1",
    "fast-autocomplete>=0.9.0",
    "humanize>=4.14.0",
    "jsonpickle>=4.1.1",
    "levenshtein>=0.27.3",
    "lxml>=6.0.2",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pyinstaller>=6.18.0",
    "pympler>=1.1",
    "pywinstyles>=1.8",
    "requests>=2.32.3",
    "requests-doh>=1.0.0",
    "ruff>=0.14.7",
    "screeninfo>=0.8.1",
    "setuptools>=80.9.0",
    "simple-toml-configurator>=1.3.0",
    "sortedcontainers>=2.4.0",
    "sqlmodel>=0.0.31",
    "tkinterdnd2>=0.4.3",
    "ttkbootstrap>=1.19.2",
]
